gcp/requirements.txt
joblib
google-cloud-storage
twilio
scikit-learn
fastapi
pandas
uvicorn
pydantic